First iteration of reducing domains
Reduce a list of domains down to their topmost domains - removing
subdomains and leaving the highest, most general domain. Hosts files
might need to specify all the subdomains, but when running your own DNS
server you don't: you just have it return NXDOMAIN for the topmost
domain. Subdomains won't be checked/resolved if a domain higher up the
chain returns NXDOMAIN.

Also this isn't actually the first iteration, but the one prior to this
one was embarrassingly naive and slow... so I'll keep that one to
myself.