Streamline reducing hosts
Remove hosts file stripping, this will be handled by a seperate
command. Auto-generate the key-indices and use that to possibly sort 63
subdomains deep. The process substitution was a remnant of the
while-read loop, we don't need it anymore: just make it pipes.

The last bit probably also means that we could change the interpreter to
/bin/sh, but I couldn't find a clear answer whether or not it supports
the -d argument to `read`. It works on my machine, but my /bin/sh is
symlinked to /bin/bash. The manpage of `sh` doesn't mention the -d
argument, so I'll let this use bash for now.