Parse blocklists into a stream of domains
Pave over the differences in hosts files and files just containing a
list of domains. This will output a list of domains, seperated by
newlines. This serves as the starting point for the rest of the scripts.

It might not be nice to run `pre` twice over the same set of files, but
this looks the cleanest. I will experiment more with it.