Tighten up the 'www' regexp and update readme
Improve the simplistic and overly eager regexp; the problem was that
when you had a domain like 'www.example' it would get reduced to just
the top-level domain. Now it only removes the 'www' domain (or multiples
thereof) if it is not directly under a top-level domain, ie the rest
contains at least 1 dot.

Update the readme with new numbers conform this updated regexp.