Fix race condition in `map-idn-conditionally.awk`
In piping the print statement into an external command I was
inadvertently introducing a race condition. It seems like the rest of
the pipeline continues whilst `idn2` is executing: everything after the
bar in the print statement is executed concurrently, while awk goes on
with the rest of the input. This causes the output of idn2, on stdout,
to clobber the stdout output of the continuing awk script.

How we call the external command now doesn't seem to have that issue. By
reading the output of the idn2 command into a variable the execution
appears to happen inline and awk doesn't continue executing.

Also update the readme, since the race did cause the reduction of
subdomains of an incompletely printed domain.