Speed up awk script in `domains-reduce.sh`
From the slowest part of the pipeline to near negligible. The speed up
provided by switching from a regular expression to basically a substring
check is fantastic. The shell script is in total more than twice as fast
now, and the limiting factor is no longer the awk script but the prior
`sort`. So further enhancements will have to focus on moving the sort into
the awk script itself and see if that's faster or not.

Also update the readme with these findings.