Script to show (sub)domain statistics
In normal operation (`d` > 0) this will count the amount of subdomains a
domain has, where `d` determines how many domains should be left at the
end. This because showing how many subdomains a top-level domain has is
not informative, usually. Hence we default to 2, but it's user
configurable by invoking the awk script with the -v option.

When invoked with a negative value for `d` it will count how the times a
domain occurs.