Add front end script `distill.sh`
Manually composing the pipeline was getting verbose and unwieldy... and
in the end it was always the plan to present a more friendly interface
to the casual user. So this is a script that brings all the parts,
required and optional, of the pipeline together - controlled by
commandline switches.

The name of the script is open to improvement, but I find that "distill"
describes the process nicely... in broad, general terms.