Change whitelist script from shell to awk
Make it the awk script it always wanted to be, the shell bits were just
for checking arguments. And with the latest additions the awk part was
really dwarfing the shell part - it started looking silly.

But at the same time this shows that using awk is kinda awkward (mostly
the "trickery" to get the script name) when dealing with the commandline
and arguments directly.