Add limited wildcard support to whitelisting
Like noted in the comments this is quite a bit slower than putting a
`grep --invert-match` into the pipeline, but that would mean splitting
the whitelisting duties in two. Not a fan of that, but if this proves to
be a speed bottleneck later on I know which way to go.