Pull grep out of `remove-whitelisted.awk`
And into a symbiotic companion script. It's definitely more involved
like this, but it's faster for most cases. Yes, it has slipped an
unavoidable (but functionality-wise superfluous) `cat` into the pipeline
when we have no wildcard domains - but that doesn't seem to impact
performance.

Having to mess with named pipes and `eval` is less than ideal, but also
a learning experience: how to dynamically communicate from another
script in the pipeline what the next step in the pipeline should
be. Using eval does open us up to code injection, but I consider the
risk minor, since the whitelist that contains these wildcard domains
should be under the control of the user. Not like the blocklists, which
come from various sources - sometimes of dubious quality - and are all
but impossible to personally curate due their size.

But I might add some countermeasures in the future. The most obvious one
would be to remove single quotes (') from wildcard domains (which are
illegal in domains anyway).